shared SandBoxedProcessor: Return SaveFileResult, remove defines
Not entirely ideal as SandboxedProcessor does not save anything,
but an improvement nevertheless over the current mess

+++ cli/main.cpp
@@ -98,7 +98,7 @@ int main(int argc, char *argv[])
 
 		QString file = args.at(1);
 		SandboxedProcessor processor(file);
-		return processor.process();
+		return static_cast<int>(processor.process());
 	}
 	Command *cmd = commandFromName(commandName, dbService);
 	if(cmd != nullptr)
